Extracted stand-alone component extending Link for the LRS action link. Preparing for #84
